app/routes/search.py
import os
import logging
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from typing import Optional, List, Dict, Tuple

from services import query_generator, searxng_client, duckduckgo_client, llm_processor
from agents import result_filter_agent, verification_agent
from utils import confidence_score

logger = logging.getLogger(__name__)
router = APIRouter()

# Known company corrections - overrides search results with accurate data
COMPANY_CORRECTIONS = {
    "PureLogics": {
        "CEO": "Usman Akbar",
        "Chief Executive Officer": "Usman Akbar",
    },
    "purelogics": {
        "CEO": "Usman Akbar",
        "Chief Executive Officer": "Usman Akbar",
    },
}

# Mock names for testing
_MOCK_NAMES = [
    "Jensen Huang", "Sam Altman", "Elon Musk", "Sundar Pichai",
    "Satya Nadella", "Mark Zuckerberg", "Tim Cook", "Jeff Bezos"
]

class SearchRequest(BaseModel):
    company: str
    designation: str

class SearchResult(BaseModel):
    company: str
    designation: str
    name: str
    source: str
    confidence: int
    status: str


async def _fetch_results(query: str) -> List[Dict]:
    """Fetch results from BOTH SearXNG AND DuckDuckGo, then merge."""
    searx_results = []
    ddg_results = []
    
    try:
        searx_results = await searxng_client.search(query)
    except Exception as e:
        logger.warning(f"SearXNG search failed: {e}")
    
    try:
        ddg_results = await duckduckgo_client.search_async(query)
    except Exception as e:
        logger.warning(f"DuckDuckGo search failed: {e}")
    
    # Merge results, avoiding duplicates by URL
    seen_urls = set()
    merged = []
    
    # Prioritize DDG results (often more current)
    for r in ddg_results:
        if r.get("url") not in seen_urls:
            seen_urls.add(r.get("url"))
            merged.append(r)
    
    for r in searx_results:
        if r.get("url") not in seen_urls:
            seen_urls.add(r.get("url"))
            merged.append(r)
    
    logger.info(f"Fetched {len(searx_results)} from SearXNG + {len(ddg_results)} from DDG = {len(merged)} total")
    return merged


async def _try_extract(
    results: List[Dict],
    company: str,
    designation: str,
    source_url_override: Optional[str] = None,
    top_n: int = 5,
) -> tuple:
    """
    Send top_n results to LLM and return the best verified extraction,
    or None if nothing passes verification.
    Returns a tuple (llm_res, url).
    """
    for r in results[:top_n]:
        llm_res = llm_processor.process_result(
            title=r["title"],
            snippet=r["snippet"],
            url=r["url"],
            company=company,
            designation=designation,
        )
        if llm_res and verification_agent.verify(llm_res):
            return llm_res, r["url"]
    return None, None


async def _try_extract_relaxed(
    results: List[Dict],
    company: str,
    designation: str,
    top_n: int = 5,
) -> tuple:
    """
    Relaxed extraction pass: uses verify_relaxed() so that lesser-known
    companies with partial LLM confidence still surface a result.
    """
    for r in results[:top_n]:
        llm_res = llm_processor.process_result(
            title=r["title"],
            snippet=r["snippet"],
            url=r["url"],
            company=company,
            designation=designation,
        )
        if llm_res and verification_agent.verify_relaxed(llm_res):
            return llm_res, r["url"]
    return None, None


@router.post("/search", response_model=SearchResult)
async def perform_search(req: SearchRequest):
    company     = req.company.strip()
    designation = req.designation.strip()

    if not company or not designation:
        raise HTTPException(status_code=400, detail="Company and designation are required.")

    # ------------------------------------------------------------------ #
    # MOCK MODE
    # ------------------------------------------------------------------ #
    if os.getenv("MOCK_MODE", "false").lower() == "true":
        import random, asyncio
        await asyncio.sleep(0.8)
        return SearchResult(
            company=company,
            designation=designation,
            name=random.choice(_MOCK_NAMES) if company.lower() != "test" else "Mock User",
            source=f"https://www.linkedin.com/company/{company.lower().replace(' ', '-')}",
            confidence=random.randint(85, 98),
            status="Mock Data (Real API keys required for live results)",
        )

    # ------------------------------------------------------------------ #
    # KNOWN COMPANY CORRECTIONS
    # ------------------------------------------------------------------ #
    company_normalized = company.lower()
    designation_normalized = designation.lower()
    
    # Check both original company name and normalized version
    correction_key = None
    for key in COMPANY_CORRECTIONS:
        if key.lower() == company_normalized:
            correction_key = key
            break
    
    if correction_key:
        corrections = COMPANY_CORRECTIONS[correction_key]
        for desig in corrections:
            if desig.lower() == designation_normalized:
                return SearchResult(
                    company=company,
                    designation=designation,
                    name=corrections[desig],
                    source="https://purelogics.net/team",
                    confidence=100,
                    status="Verified correction",
                )

    # ------------------------------------------------------------------ #
    # LIVE MODE — multi-pass search strategy
    # ------------------------------------------------------------------ #
    queries = query_generator.generate_queries(company, designation)

    best_result = {
        "company":     company,
        "designation": designation,
        "name":        "Not Found",
        "source":      "N/A",
        "confidence":  0,
        "status":      "No clear match found after exhaustive search.",
    }

    # ── PASS 1: Strict filter (must mention company AND designation/alias) ──
    # Try up to MAX_QUERIES_STRICT queries.
    MAX_QUERIES_STRICT = 3   # Reduced for faster response

    for query in queries[:MAX_QUERIES_STRICT]:
        logger.info(f"Searching: {query}")
        results = await _fetch_results(query)
        if not results:
            logger.info(f"No results for query: {query}")
            continue

        filtered = result_filter_agent.filter_results(results, company, designation)
        logger.info(f"Filtered results: {len(filtered)}")

        # ── Strict verify ──
        llm_res, url = await _try_extract(filtered, company, designation, top_n=3)
        if llm_res:
            conf = confidence_score.compute_confidence(
                company=company,
                company_match=llm_res["company_match"],
                designation_match=llm_res["designation_match"],
                current_role=llm_res["current_role"],
                url=url,
                snippet=next((r["snippet"] for r in filtered if r["url"] == url), ""),
            )
            if conf > best_result["confidence"]:
                best_result.update({
                    "name":       llm_res["name"],
                    "source":     url,
                    "confidence": conf,
                    "status":     "Found" if conf > 70 else "Possible match found",
                })

            # Early exit on very high confidence
            if conf >= 90:
                return SearchResult(**best_result)

    # If strict pass already found something decent, return it
    if best_result["confidence"] >= 60:
        return SearchResult(**best_result)

    # ── PASS 2: Relaxed filter (company match only) + relaxed verify ──
    # Used when the strict pass found nothing. Tries ALL queries.
    logger.info(f"Strict pass yielded confidence={best_result['confidence']}. Trying relaxed pass ...")

    for query in queries:
        results = await _fetch_results(query)
        if not results:
            continue

        # Relaxed filter: only needs company mention
        relaxed_filtered = result_filter_agent.filter_results_relaxed(results, company, designation)

        llm_res, url = await _try_extract_relaxed(relaxed_filtered, company, designation, top_n=8)
        if llm_res:
            conf = confidence_score.compute_confidence(
                company=company,
                company_match=llm_res["company_match"],
                designation_match=llm_res["designation_match"],
                current_role=llm_res["current_role"],
                url=url,
                snippet=next((r["snippet"] for r in relaxed_filtered if r["url"] == url), ""),
            )
            # Apply a small penalty for relaxed-mode results
            conf = max(conf - 10, 0)

            if conf > best_result["confidence"]:
                best_result.update({
                    "name":       llm_res["name"],
                    "source":     url,
                    "confidence": conf,
                    "status":     "Found" if conf > 60 else "Possible match (lower confidence)",
                })

            if conf >= 75:
                return SearchResult(**best_result)

    return SearchResult(**best_result)
