app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
groq==0.5.0
python-multipart==0.0.9
pandas==2.2.1
aiofiles==23.2.1
duckduckgo-search==5.3.1b1
slowapi==0.1.9
